+++ apps/api-frontend/src/app/pages/search-result.css
@@ -1,0 +1,11 @@
+.loading-container {
+  display: flex;
+  width: 100%;
+  align-items: center;
+  justify-content: center;
+}
+
+.single-result-container p {
+  padding: 1rem;
+  text-align: left;
+}
