--- apps/api-frontend/src/app/pages/search.css
@@ -1,4 +1,0 @@
-.single-result-container p {
-  padding: 1rem;
-  text-align: left;
-}
