+++ package.json
@@ -35,12 +35,15 @@
     "fs": "0.0.1-security",
     "natural": "^2.1.5",
     "node-fetch": "^2.6.0",
+    "promise-hook": "^0.1.4",
     "react": "16.12.0",
     "react-collapsible": "^2.8.0",
     "react-dom": "16.12.0",
     "react-motion": "^0.5.2",
+    "react-router": "^5.2.0",
     "react-router-dom": "5.1.2",
     "react-search-box": "^2.0.2",
+    "react-spinners": "^0.9.0",
     "react-table": "^7.2.0",
     "synonyms": "^1.0.1",
     "util": "^0.12.3",
